# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tracer77 API",
        "description": "API d'authentification et de géolocalisation pour Tracer77",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.1.77:8000/",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "summary": "Inscription d'un nouvel utilisateur",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Daryl"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "daryl@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inscription réussie"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "Connexion d'un utilisateur",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "daryl@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie"
                    },
                    "401": {
                        "description": "Email ou mot de passe incorrect"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "summary": "Déconnexion de l'utilisateur",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "summary": "Récupérer l'utilisateur connecté",
                "operationId": "e5478e84ace512f8d9653dd7e920ea2a",
                "responses": {
                    "200": {
                        "description": "Utilisateur retourné"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/update-name": {
            "put": {
                "summary": "Modifier le nom",
                "operationId": "91eed3a5941cde6e3820e95fc019191e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Daryl"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nom mis à jour"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/update-password": {
            "put": {
                "summary": "Modifier le mot de passe",
                "operationId": "912b71f24278e89f26f6c0bca8e9e730",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "example": "nouveau123"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string",
                                        "example": "nouveau123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe mis à jour"
                    },
                    "401": {
                        "description": "Mot de passe actuel incorrect"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/delete": {
            "delete": {
                "summary": "Supprimer le compte",
                "operationId": "f087c70fa14255f10a79bd7fd3279964",
                "responses": {
                    "200": {
                        "description": "Compte supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/forgot-password": {
            "post": {
                "summary": "Demander un code de réinitialisation",
                "operationId": "7af8817e88ba5874543ff38f188248f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "daryl@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code envoyé si l'email existe"
                    }
                }
            }
        },
        "/api/reset-password": {
            "post": {
                "summary": "Réinitialiser le mot de passe avec le code reçu",
                "operationId": "00c0aad74830221c6da409a1355ec736",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "daryl@example.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "le_code_recu_par_email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "nouveaupass123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "nouveaupass123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe réinitialisé"
                    },
                    "422": {
                        "description": "Code invalide ou expiré"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez votre token Sanctum ici",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}